Initial prompt setup
To open a project, first define PROJECT_NAME to be what you want, then
source open-project.sh

It will check out the manifest files that match your project name
Once done, it will set up the prompt with the name of the project. If
you above the project root directory or at it the prompt just says the
name. If you enter any sub repository, it will add the name of this
repository, and it will be fixed no matter how many directories you go
down